Add awaiting of ratelimit refresh. Read GIHUB_TOKEN from global()(strange but it's name not available in function).  Remove all dublicate of request with limit fuction
